Support http bodies up to 2^64-1 bytes large.

+++ data-plane/http-common/include/ESHttpClientStream.h
@@ -30,8 +30,8 @@ class HttpClientStream : public HttpStream {
    * @return ESB_SUCCESS if 1+ bytes consumed, ESB_AGAIN if send buffer full and
    * underlying socket send buffer is full, another error code otherwise.
    */
-  virtual ESB::Error sendRequestBody(unsigned const char *body, ESB::UInt32 bytesOffered,
-                                     ESB::UInt32 *bytesConsumed) = 0;
+  virtual ESB::Error sendRequestBody(unsigned const char *body, ESB::UInt64 bytesOffered,
+                                     ESB::UInt64 *bytesConsumed) = 0;
 
   /**
    * Determine how many bytes of response body data can be read without
@@ -44,7 +44,7 @@ class HttpClientStream : public HttpStream {
    * empty and there is no data to be read on the underlying socket, or another
    * error code otherwise.
    */
-  virtual ESB::Error responseBodyAvailable(ESB::UInt32 *bytesAvailable) = 0;
+  virtual ESB::Error responseBodyAvailable(ESB::UInt64 *bytesAvailable) = 0;
 
   /**
    * Read up to bytesRequested of response body data.
@@ -57,7 +57,7 @@ class HttpClientStream : public HttpStream {
    * returned, ESB_INVALID_ARGUMENT if bytesRequested exceeds bytesAvailable (use requestBodyAvailable() first), another
    * error code otherwise.
    */
-  virtual ESB::Error readResponseBody(unsigned char *body, ESB::UInt32 bytesRequested, ESB::UInt32 *bytesRead) = 0;
+  virtual ESB::Error readResponseBody(unsigned char *body, ESB::UInt64 bytesRequested, ESB::UInt64 *bytesRead) = 0;
 
  private:
   // Disabled
