+++ data-plane/http1/include/ESHttpClientSocket.h
@@ -157,9 +157,9 @@ class HttpClientSocket : public HttpSocket, public HttpClientStream {
 
   virtual bool secure() const;
 
-  virtual ESB::Error sendRequestBody(unsigned const char *chunk, ESB::UInt32 bytesOffered, ESB::UInt32 *bytesConsumed);
-  virtual ESB::Error responseBodyAvailable(ESB::UInt32 *bytesAvailable);
-  virtual ESB::Error readResponseBody(unsigned char *chunk, ESB::UInt32 bytesRequested, ESB::UInt32 *bytesRead);
+  virtual ESB::Error sendRequestBody(unsigned const char *chunk, ESB::UInt64 bytesOffered, ESB::UInt64 *bytesConsumed);
+  virtual ESB::Error responseBodyAvailable(ESB::UInt64 *bytesAvailable);
+  virtual ESB::Error readResponseBody(unsigned char *chunk, ESB::UInt64 bytesRequested, ESB::UInt64 *bytesRead);
 
   //
   // ESB::EmbeddedMapElement (for connection pool lookups)
@@ -241,15 +241,15 @@ class HttpClientSocket : public HttpSocket, public HttpClientStream {
     _state &= ~flag;
   }
 
-  ESB::Error currentChunkBytesAvailable(ESB::UInt32 *bytesAvailable);
-  ESB::Error formatStartChunk(ESB::UInt32 chunkSize, ESB::UInt32 *maxChunkSize);
+  ESB::Error currentChunkBytesAvailable(ESB::UInt64 *bytesAvailable);
+  ESB::Error formatStartChunk(ESB::UInt64 chunkSize, ESB::UInt64 *maxChunkSize);
   ESB::Error formatEndChunk();
   ESB::Error fillReceiveBuffer();
   ESB::Error flushSendBuffer();
 
   int _state;
-  ESB::UInt32 _bodyBytesWritten;
-  ESB::UInt32 _bytesAvailable;
+  ESB::UInt64 _bodyBytesWritten;
+  ESB::UInt64 _bytesAvailable;
   HttpMultiplexerExtended &_multiplexer;
   HttpClientHandler &_handler;
   HttpClientTransaction *_transaction;
