+++ data-plane/multiplexers/source/ESHttpProxyMultiplexer.cpp
@@ -21,20 +21,20 @@ class HttpNullClientHandler : public HttpClientHandler {
   }
 
   virtual ESB::Error offerRequestBody(HttpMultiplexer &multiplexer, HttpClientStream &stream,
-                                      ESB::UInt32 *bytesAvailable) {
+                                      ESB::UInt64 *bytesAvailable) {
     assert(0 == "HttpNullClientHandler called");
     return ESB_NOT_IMPLEMENTED;
   }
 
   virtual ESB::Error produceRequestBody(HttpMultiplexer &multiplexer, HttpClientStream &stream, unsigned char *chunk,
-                                        ESB::UInt32 bytesRequested) {
+                                        ESB::UInt64 bytesRequested) {
     assert(0 == "HttpNullClientHandler called");
     return ESB_NOT_IMPLEMENTED;
   }
 
   virtual ESB::Error consumeResponseBody(HttpMultiplexer &multiplexer, HttpClientStream &stream,
-                                         const unsigned char *chunk, ESB::UInt32 chunkSize,
-                                         ESB::UInt32 *bytesConsumed) {
+                                         const unsigned char *chunk, ESB::UInt64 chunkSize,
+                                         ESB::UInt64 *bytesConsumed) {
     assert(0 == "HttpNullClientHandler called");
     return ESB_NOT_IMPLEMENTED;
   }
@@ -67,19 +67,19 @@ class HttpNullServerHandler : public HttpServerHandler {
   }
 
   virtual ESB::Error consumeRequestBody(HttpMultiplexer &multiplexer, HttpServerStream &stream,
-                                        unsigned const char *chunk, ESB::UInt32 chunkSize, ESB::UInt32 *bytesConsumed) {
+                                        unsigned const char *chunk, ESB::UInt64 chunkSize, ESB::UInt64 *bytesConsumed) {
     assert(0 == "HttpNullServerHandler called");
     return ESB_NOT_IMPLEMENTED;
   }
 
   virtual ESB::Error offerResponseBody(HttpMultiplexer &multiplexer, HttpServerStream &stream,
-                                       ESB::UInt32 *bytesAvailable) {
+                                       ESB::UInt64 *bytesAvailable) {
     assert(0 == "HttpNullServerHandler called");
     return ESB_NOT_IMPLEMENTED;
   }
 
   virtual ESB::Error produceResponseBody(HttpMultiplexer &multiplexer, HttpServerStream &stream, unsigned char *chunk,
-                                         ESB::UInt32 bytesRequested) {
+                                         ESB::UInt64 bytesRequested) {
     assert(0 == "HttpNullServerHandler called");
     return ESB_NOT_IMPLEMENTED;
   }
